Duplicate Github Actions Workflows from heimdall_tools
Instead of tagging a new release on every merge to master, this takes advantage of GitHub Releases to release a new version.

This switches over to the git-version-bump gem (https://github.com/rbclark/git-version-bump) which uses the git tag at the time of building to determine the version. This means no more updating the version.rb on every release, meaning 33% less commits to master.

Changelog is still built on every commit and will show the currently merged but unreleased changes under "Unreleased". AFTER a tag is created, a new Changelog will be automatically pushed to master.